+++ adapter/cli/src/main/kotlin/CliMain.kt
@@ -2,7 +2,7 @@ import adapters.primary.CliAdapter
 import application.usecases.TokenizeUseCase
 
 fun main(args: Array<String>) {
-    val tokenizeUseCase = TokenizeUseCase()
-    val cliAdapter = CliAdapter(tokenizeUseCase)
+    val tokenizerPort = TokenizeUseCase()
+    val cliAdapter = CliAdapter(tokenizerPort)
     cliAdapter.handleCommand(args)
 }